MindBloomBack/requirements.txt
fastapi==0.109.0
supabase==2.3.1
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.0
uvicorn==0.27.0
tensorflow==2.18.0
numpy==1.26.0
flask==2.3.3
PILLOW==9.5.0
